+++ media/main.js
@@ -1,10 +1,5 @@
-//@ts-check
-
-// See https://github.com/microsoft/vscode-extension-samples/blob/main/webview-view-sample/media/main.js
-
-// This script will be run within the webview itself
-// It cannot access the main VS Code APIs directly.
 (function () {
+    // See https://github.com/microsoft/vscode-extension-samples/blob/main/webview-view-sample/media/main.js
     // @ts-ignore
     const vscode = acquireVsCodeApi();
 
@@ -16,6 +11,20 @@
     stopButton?.addEventListener('click', () => {
         stopHugo();
     });
+    const includeDraftsCheckbox = document.getElementById('include-drafts');
+    const includeFutureCheckbox = document.getElementById('include-future');
+    const includeExpiredCheckbox = document.getElementById('include-expired');
+
+    includeDraftsCheckbox?.addEventListener('change', () => {saveState();});
+    includeFutureCheckbox?.addEventListener('change', () => {saveState();});
+    includeExpiredCheckbox?.addEventListener('change', () => {saveState();});
+
+    const previousState = vscode.getState();
+    if (previousState) {
+        includeDraftsCheckbox.checked = previousState.drafts;
+        includeFutureCheckbox.checked = previousState.future;
+        includeExpiredCheckbox.checked = previousState.expired;
+    }
 
     // Handle messages sent from the extension to the webview
     window.addEventListener('message', event => {
@@ -24,69 +33,39 @@
             case 'hugoStarted':
                 {
                     startButton?.setAttribute('disabled', 'disabled');
+                    includeDraftsCheckbox?.setAttribute('disabled', 'disabled');
+                    includeFutureCheckbox?.setAttribute('disabled', 'disabled');
+                    includeExpiredCheckbox?.setAttribute('disabled', 'disabled');
                     stopButton?.attributes.removeNamedItem('disabled');
                     break;
                 }
             case 'hugoStopped':
                 {
                     startButton?.attributes.removeNamedItem('disabled');
+                    includeDraftsCheckbox?.attributes.removeNamedItem('disabled');
+                    includeFutureCheckbox?.attributes.removeNamedItem('disabled');
+                    includeExpiredCheckbox?.attributes.removeNamedItem('disabled');
                     stopButton?.setAttribute('disabled', 'disabled');
                     break;
                 }
         }
     });
 
-    // /**
-    //  * @param {Array<{ value: string }>} colors
-    //  */
-    // function updateColorList(colors) {
-    //     const ul = document.querySelector('.color-list');
-    //     if (!ul) {
-    //         return;
-    //     }
-    //     ul.textContent = '';
-    //     for (const color of colors) {
-    //         const li = document.createElement('li');
-    //         li.className = 'color-entry';
-
-    //         const colorPreview = document.createElement('div');
-    //         colorPreview.className = 'color-preview';
-    //         colorPreview.style.backgroundColor = `#${color.value}`;
-    //         colorPreview.addEventListener('click', () => {
-    //             onColorClicked(color.value);
-    //         });
-    //         li.appendChild(colorPreview);
-
-    //         const input = document.createElement('input');
-    //         input.className = 'color-input';
-    //         input.type = 'text';
-    //         input.value = color.value;
-    //         input.addEventListener('change', (e) => {
-    //             // @ts-ignore
-    //             const value = e.target.value;
-    //             if (!value) {
-    //                 // Treat empty value as delete
-    //                 colors.splice(colors.indexOf(color), 1);
-    //             } else {
-    //                 color.value = value;
-    //             }
-    //             updateColorList(colors);
-    //         });
-    //         li.appendChild(input);
-
-    //         ul.appendChild(li);
-    //     }
-
-    //     // Update the saved state
-    //     vscode.setState({ colors: colors });
-    // }
-
-
     function startHugo() {
-        vscode.postMessage({ type: 'startHugo' });
+        const drafts = includeDraftsCheckbox?.checked;
+        const future = includeFutureCheckbox?.checked;
+        const expired = includeExpiredCheckbox?.checked;
+        vscode.postMessage({ type: 'startHugo', drafts, future, expired});
     }
     function stopHugo() {
         vscode.postMessage({ type: 'stopHugo' });
+    }
+
+    function saveState(){
+        const drafts = includeDraftsCheckbox?.checked;
+        const future = includeFutureCheckbox?.checked;
+        const expired = includeExpiredCheckbox?.checked;
+        vscode.setState({drafts, future, expired});
     }
 }());